+++ modules/Students/Letters.php
@@ -117,11 +117,6 @@ if ( $_REQUEST['modfunc'] === 'save'
 	{
 		unset( $_ROSARIO['DrawHeader'] );
 
-		if ( $_REQUEST['mailing_labels'] === 'Y' )
-		{
-			echo '<br /><br /><br />';
-		}
-
 		if ( ! $_REQUEST['hide_headers'] )
 		{
 			DrawHeader( '&nbsp;' );
@@ -131,12 +126,13 @@ if ( $_REQUEST['modfunc'] === 'save'
 		}
 		elseif ( $_REQUEST['mailing_labels'] === 'Y' )
 		{
-			echo '<br /><br /><br /><br />';
+			echo '<br /><br /><br /><br /><br /><br /><br /><br /><br />';
 		}
 
 		if ( $_REQUEST['mailing_labels'] === 'Y' )
 		{
-			echo '<br /><br /><table class="width-100p"><tr><td style="width:50px;"> &nbsp; </td><td>' . $student['MAILING_LABEL'] . '</td></tr></table><br />';
+			// @since 11.6 Add Mailing Label position
+			echo MailingLabelPositioned( $student['MAILING_LABEL'] );
 		}
 
 		$substitutions = [